third_party/WebKit/LayoutTests/html/dialog/dialog-return-value-expected.txt

Tests that dialog.returnValue is settable and returns the last value set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dialog.returnValue is ''
PASS dialog.returnValue is 'Setting value directly'
PASS dialog.returnValue is 'null'
PASS dialog.returnValue is ''
PASS dialog.returnValue is '7'
PASS dialog.returnValue is 'Return value set from close()'
PASS dialog.returnValue is ''
PASS dialog.returnValue is 'null'
PASS dialog.returnValue is 'Should not change because no argument to close()'
PASS dialog.returnValue is 'Should not change because of undefined argument to close()'
PASS dialog.close('blah') did not throw exception.
PASS dialog.returnValue is 'Should not change because of no-op close()'
PASS successfullyParsed is true

TEST COMPLETE
